Convert all time variables to double
The differences between time types were causing crashes. Times of particles
were previously doubles and Clock class times were floats. This caused
strange numerical errors at taking differences and comparisons. SMASH
propagates particles from action to action and because of these numerical errors
next action sometimes turned out to be earlier than previous.

All the times should have consistent types - either double or float.
Making all times float would require having FourVector<float>, which seems a
harder solution to me than converting time variables to double.

+++ src/action.cc
@@ -23,7 +23,7 @@
 
 namespace Smash {
 
-Action::Action(const ParticleList &in_part, float time)
+Action::Action(const ParticleList &in_part, double time)
               : incoming_particles_(in_part),
                 time_of_execution_(time+in_part[0].position().x0()) {}
